build.sh: Fix error if build_results directory not found + update .gitignore

+++ build.sh
@@ -113,7 +113,10 @@ fi
  echo ""
  echo ""
  echo "		==> [WORKING] moving files!"
+ 
+ if [ "$(ls -A ./build_results)" ]; then
  rm -R ./build_results
+ fi
  mkdir build_results
  cp out/target/product/${TARGET_PRODUCT}/boot.img build_results
  cp out/target/product/${TARGET_PRODUCT}/kernel build_results
